+++ www/act-map.js
@@ -123,6 +123,8 @@ act('setSignOut', setSignOut);
 act('setMail', setMail);
 act('setPwGo', setPwGo);
 act('obSkipDraw', obSkipDraw);
+act('obTourDone', obTourDone);
+act('obTourNext', obTourNext);
 act('openSnd', openSnd);
 act('openSndAdd', openSndAdd);
 act('ltGo', ltGo);